requirements.txt
flask>=3.1.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
flask-wtf>=1.2.0
flask-migrate>=4.0.0
email-validator>=2.1.0
pillow>=10.0.0
werkzeug>=3.1.0
